Handle CLI arguments in `HighWatermarkCommand`
All of the commands that make use of the `HighWatermarkCommand` base
class accept a very similar set of command line arguments, with only
minor variations. The only differences are:

- Only `memray flamegraph` supports `--temporal-leaks`
- Only `memray flamegraph` supports `--split-threads`
- Only `memray transform` has a required positional argument

Set up the `argparse` parser in the base class. Make it choose whether
or not to add the `--temporal-leaks` argument based on whether or not
a `temporal_reporter_factory` was provided. Extend the base class
implementation for `memray flamegraph` and `memray transform` to add
support for the other reporter-specific arguments.
