+++ src/memray/commands/common.py
@@ -3,6 +3,7 @@ import os
 import pathlib
 import sys
 from pathlib import Path
+from textwrap import dedent
 from typing import Any
 from typing import Iterable
 from typing import Optional
@@ -176,6 +177,65 @@ class HighWatermarkCommand:
                 show_memory_leaks=show_memory_leaks,
                 merge_threads=merge_threads,
             )
+
+    def prepare_parser(self, parser: argparse.ArgumentParser) -> None:
+        parser.add_argument(
+            "-o",
+            "--output",
+            help="Output file name",
+            default=None,
+        )
+        parser.add_argument(
+            "-f",
+            "--force",
+            help="If the output file already exists, overwrite it",
+            action="store_true",
+            default=False,
+        )
+        alloc_type_group = parser.add_mutually_exclusive_group()
+        alloc_type_group.add_argument(
+            "--leaks",
+            help="Show memory leaks, instead of peak memory usage",
+            action="store_true",
+            dest="show_memory_leaks",
+            default=False,
+        )
+        if self.temporal_reporter_factory:
+            alloc_type_group.add_argument(
+                "--temporal-leaks",
+                help=(
+                    "Generate a dynamic flame graph showing allocations performed"
+                    " in a user-selected time range and not freed before the end"
+                    " of that time range."
+                ),
+                action="store_true",
+                default=False,
+            )
+        alloc_type_group.add_argument(
+            "--temporary-allocation-threshold",
+            metavar="N",
+            help=dedent(
+                """
+                Report temporary allocations, as opposed to leaked allocations
+                or high watermark allocations.  An allocation is considered
+                temporary if at most N other allocations occur before it is
+                deallocated.  With N=0, an allocation is temporary only if it
+                is immediately deallocated before any other allocation occurs.
+                """
+            ),
+            action="store",
+            dest="temporary_allocation_threshold",
+            type=int,
+            default=-1,
+        )
+        alloc_type_group.add_argument(
+            "--temporary-allocations",
+            help="Equivalent to --temporary-allocation-threshold=1",
+            action="store_const",
+            dest="temporary_allocation_threshold",
+            const=1,
+        )
+        parser.add_argument("results", help="Results of the tracker run")
 
     def run(
         self, args: argparse.Namespace, parser: argparse.ArgumentParser, **kwargs: Any
